Update script.js
trying a new thing with this request stuff. for now i am changing the name of the constructor to "Requester" because to be honest, i am not sure how much this constructor really contributes to APIs. i am trying a new approach to connecting to different branches within the "base" URL. each new branch is now defined by creating a new object and setting the appropriate properties. instead of having a singular XMLHttpRequest existing, each branch object gets its own XMLHttpRequest which means that while for all i know performance may not be as good, requests can happen simultaneously. if you want a callback that waits for every request to be ready, you can use onready so every response is ready before anything happens. i've temporarily removed the lines of code that involve HTTP headers and REST